(wip): UsersController

+++ Servidor-Estatico/controllers/userController.js
@@ -2,8 +2,10 @@
  * Tiene como responsabilidad el manejo HTTP de los usuarios
  */
 
-import { isValidUser } from "../lib/validations.js"
+import { fileExists } from "../lib/data.js"
+import { isValidUser, isValidUser as userWithRequiredParams } from "../lib/validations.js"
 import { UserModel } from "../models/userModel.js"
+import * as path from 'node:path'
 
 export const userController = async (req, res, payloadEnBruto, urlParts) => {
   /**
@@ -37,7 +39,7 @@ export const userController = async (req, res, payloadEnBruto, urlParts) => {
   else if (req.method == 'POST' && payloadEnBruto) {
     try {
       let user = JSON.parse(payloadEnBruto)
-      if(!isValidUser(user)) throw new Error('Usuario Inválido');
+      if(!userWithRequiredParams(user)) throw new Error('Usuario Inválido');
 
       let fueCreado = await UserModel.create(user.telefono, user)
       if(fueCreado) {
@@ -60,15 +62,60 @@ export const userController = async (req, res, payloadEnBruto, urlParts) => {
    * DELETE /api/users/id
    */
   else if (req.method == 'DELETE' && urlParts[2]) {
-    // TODO
-    // FIXME: No borra usuarios
-    await UserModel.delete(urlParts[2])
-
-    res.writeHead(200, 'OK', { "content-type": "application/json" })
-    res.end(JSON.stringify({ status: 'deleted' }))
+    try {
+      let userExists =  await UserModel.exists(urlParts[2])
+      if(userExists) {
+        await UserModel.delete(urlParts[2])
+        res.writeHead(200, 'OK', { "content-type": "application/json" })
+        res.end(JSON.stringify({ status: 'deleted' }))
+      } else {
+        res.writeHead(404, 'Not Found', { "content-type": "text/plain" })
+        res.end('Usuario no encontrado')
+      }
+    } catch (err) {
+      console.error("Error eliminando usuario", err)
+      res.writeHead(500, 'Internal Server Error', { "content-type": "text/plain" })
+      res.end('Error de servidor')
+    }
   }
 
   /**
    * PUT /api/users/id  + payload
    */
+  else if (req.method == 'PUT' && urlParts[2] && payloadEnBruto) {
+    try {
+      const payload = JSON.parse(payloadEnBruto)
+      const user = await UserModel.getById(urlParts[2])
+      if(user) {
+        /**
+         * Validamos Payload
+         */
+        if(isValidUser(payload)) {
+          delete payload.telefono
+          let payloadKeys = Object.keys(payload)
+          if(payloadKeys.length === 0) throw new Error('Payload Inválido');
+          let usuarioActualizado = { ...user, ...payload }
+
+          try {
+            await UserModel.update(urlParts[2], usuarioActualizado)
+
+            res.writeHead(200, 'OK', { "content-type": "text/plain" })
+            res.end("Usuario Actualizado")
+          } catch (err) {
+            res.writeHead(500, "Internal Server Error", { "content-type": "text/plain" })
+            res.end("Error Interno de Servidor")
+          }
+        } else {
+          throw new Error('Payload Inválido')
+        }
+      } else {
+        res.writeHead(404, 'Not Found', { "content-type": "text/plain" })
+        res.end("Usuario no encontrado")
+      }
+    } catch (err) {
+      console.error(err)
+      res.writeHead(400, 'Bad Request', { "content-type": "text/plain" })
+      res.end("Solicitud mal hecha")
+    }
+  }
 }
